+++ agent/src/it/transplant-indices/app/src/test/kotlin/net/onedaybeard/graftt/IntegrationTests.kt
@@ -1,0 +1,21 @@
+package net.onedaybeard.graftt
+
+import org.junit.Test
+
+class IntegrationTests {
+
+    @Test
+    fun `verify foo transplant`() {
+        Foo.foo()
+    }
+
+    @Test
+    fun `verify bar transplant`() {
+        Bar.bar()
+    }
+
+    @Test
+    fun `verify bar2 transplant`() {
+        Bar2.bar2()
+    }
+}