+++ agent/src/it/transplant-indices/bar-transplants/src/main/java/net/onedaybeard/graftt/Bar2Transplant.java
@@ -1,0 +1,10 @@
+package net.onedaybeard.graftt;
+
+
+@Graft.Recipient(Bar2.class)
+public class Bar2Transplant {
+    @Graft.Fuse
+    public static void bar2() {
+        // ok
+    }
+}