+++ agent/src/it/transplant-indices/bar-transplants/src/main/java/net/onedaybeard/graftt/BarTransplant.java
@@ -1,0 +1,10 @@
+package net.onedaybeard.graftt;
+
+
+@Graft.Recipient(Bar.class)
+public class BarTransplant {
+    @Graft.Fuse
+    public static void bar() {
+        // ok
+    }
+}