+++ agent/src/it/transplant-indices/base/src/main/java/net/onedaybeard/graftt/Bar.java
@@ -1,0 +1,7 @@
+package net.onedaybeard.graftt;
+
+public class Bar {
+    public static void bar() {
+        throw new IllegalStateException("bar not replaced by transplant");
+    }
+}