+++ agent/src/it/transplant-indices/base/src/main/java/net/onedaybeard/graftt/Bar2.java
@@ -1,0 +1,7 @@
+package net.onedaybeard.graftt;
+
+public class Bar2 {
+    public static void bar2() {
+        throw new IllegalStateException("bar2 not replaced by transplant");
+    }
+}