+++ agent/src/it/transplant-indices/base/src/main/java/net/onedaybeard/graftt/Foo.java
@@ -1,0 +1,7 @@
+package net.onedaybeard.graftt;
+
+public class Foo {
+    public static void foo() {
+        throw new IllegalStateException("foo not replaced by transplant");
+    }
+}