+++ agent/src/it/transplant-indices/foo-transplants/src/main/java/net/onedaybeard/graftt/FooTransplant.java
@@ -1,0 +1,10 @@
+package net.onedaybeard.graftt;
+
+
+@Graft.Recipient(Foo.class)
+public class FooTransplant {
+    @Graft.Fuse
+    public static void foo() {
+        // ok
+    }
+}