+++ graftt-maven-plugin/src/it/generate-index/src/main/java/net/onedaybeard/graftt/BarTransplant.java
@@ -1,0 +1,6 @@
+package net.onedaybeard.graftt;
+
+@Graft.Recipient(Bar.class)
+public class BarTransplant {
+
+}