+++ graftt-maven-plugin/src/it/generate-index/src/main/java/net/onedaybeard/graftt/FooTransplant.java
@@ -1,0 +1,6 @@
+package net.onedaybeard.graftt;
+
+@Graft.Recipient(Foo.class)
+public class FooTransplant {
+
+}